pyproject.toml
[tool.poetry]
name = "lam-ai"
version = "1.0.0"
description = "Proposal Generation"
authors = ["manthan <manthanpatel898@gmail.com>"]
license = "Bytes"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
flask-jwt-extended = "^4.6.0"
pymongo = "^4.8.0"
python-dotenv = "^1.0.1"
passlib = "^1.7.4"
flask-pymongo = "^2.3.0"
flask-smorest = "^0.44.0"
flask-cors = "^5.0.0"
stripe = "^10.11.0"
schemas = "^0.7.1"
bcrypt = "^4.2.0"
cryptography = "^43.0.1"
openai = "0.28"
weasyprint = "^62.3"
pandas = "^2.2.3"
openpyxl = "^3.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
